requirements.txt
Flask
nltk
sentence-transformers
openai==0.27.3
pandas
scikit-learn
numpy
requests
python-dotenv
gunicorn